refactor: rename useful.py to mock_language_model.py and update imports
